requirements.txt
fastapi
uvicorn
paho-mqtt
jinja2
python-multipart